+++ src/metric/definition_test.go
@@ -73,6 +73,14 @@ func TestToNumericBoolean(t *testing.T) {
 	assert.Equal(t, 0, v)
 	assert.NoError(t, err)
 
+	v, err = toNumericBoolean("True")
+	assert.Equal(t, 1, v)
+	assert.NoError(t, err)
+
+	v, err = toNumericBoolean("False")
+	assert.Equal(t, 0, v)
+	assert.NoError(t, err)
+
 	v, err = toNumericBoolean("invalid")
 	assert.Nil(t, v)
 	assert.EqualError(t, err, "value can not be converted to numeric boolean")